fix(test): move RecordingMsOrder to case-stable PSR-4 path
Linux CI could not autoload MiniShop3\Tests\Support from tests/support
(case mismatch). Place the stub at tests/RecordingMsOrder.php instead.
