+++ core/components/minishop3/tests/Integration/Order/OrderDraftCostClampTest.php
@@ -7,7 +7,7 @@ namespace MiniShop3\Tests\Integration\Order;
 use MiniShop3\MiniShop3;
 use MiniShop3\Services\Order\OrderDraftManager;
 use MiniShop3\Services\Order\OrderService;
-use MiniShop3\Tests\Support\RecordingMsOrder;
+use MiniShop3\Tests\RecordingMsOrder;
 use MODX\Revolution\modX;
 use PHPUnit\Framework\TestCase;
 
@@ -23,7 +23,6 @@ final class OrderDraftCostClampTest extends TestCase
     {
         if (!class_exists(modX::class, false)) {
             require_once dirname(__DIR__, 2) . '/stubs/ModxStub.php';
-        require_once dirname(__DIR__, 2) . '/support/RecordingMsOrder.php';
         }
 
         $orderService = new OrderService(new modX());
